+++ doc/versionHistory.rst
@@ -13,6 +13,7 @@ Version History
 -------------
 
 * Fix test data by adjusting ROTPA values in headers after DM-49838.
+* Enforce additional possible single threading options.
 
 .. _lsst.ts.wep-14.1.1: